refactor: seperated getMostPopularApps into lib folder
